requirements.txt
transformers 
datasets 
faiss-cpu 
nltk 
torch
scikit-learn
git+https://github.com/huggingface/accelerate